+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/KotlinAndroidExtension.kt
@@ -2,6 +2,7 @@ package com.javiersc.hubdle.extensions.kotlin.android
 
 import com.javiersc.hubdle.extensions.HubdleDslMarker
 import com.javiersc.hubdle.extensions._internal.PluginIds
+import com.javiersc.hubdle.extensions.kotlin.android.application.KotlinAndroidApplicationExtension
 import com.javiersc.hubdle.extensions.kotlin.android.library.KotlinAndroidLibraryExtension
 import javax.inject.Inject
 import org.gradle.api.Action
@@ -10,11 +11,19 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
 
 @HubdleDslMarker
-public open class KotlinAndroidExtension
-@Inject
-constructor(
-    objects: ObjectFactory,
-) {
+public open class KotlinAndroidExtension @Inject constructor(objects: ObjectFactory) {
+
+    public val application: KotlinAndroidApplicationExtension = objects.newInstance()
+
+    @HubdleDslMarker
+    public fun Project.application(
+        action: Action<in KotlinAndroidApplicationExtension> = Action {}
+    ) {
+        pluginManager.apply(PluginIds.Android.application)
+        pluginManager.apply(PluginIds.Android.kotlin)
+        application.run { isEnabled = true }
+        action.execute(application)
+    }
 
     public val library: KotlinAndroidLibraryExtension = objects.newInstance()